chore: separate files home
