+++ next.config.js
@@ -2,14 +2,6 @@
 const nextConfig = {
   reactStrictMode: true,
   swcMinify: true,
-  pageExtensions: [
-    'page.tsx',
-    'page.ts',
-    'page.jsx',
-    'page.js',
-    'index.tsx',
-    'index.jsx',
-  ],
 }
 
 module.exports = nextConfig
