+++ pages/index.tsx
@@ -1,14 +1,15 @@
 import ArrowRightAltIcon from '@mui/icons-material/ArrowRightAlt'
+import type { NextPage } from 'next'
 import { useEffect } from 'react'
 import { useGlitch } from 'react-powerglitch'
 import Sound from 'react-sound'
 
 import Equalizer from '../src/components/Equalizer'
+import * as S from '../src/styles/stylesHome'
 import { SoundClickButton, SoundGlitch } from '../src/utils/Sounds'
-import { settingsSound } from './settings'
-import * as S from './styles'
+import { settingsSound } from '../src/utils/settingsHome'
 
-const Home = () => {
+const Home: NextPage = () => {
   const glitch = useGlitch()
 
   useEffect(() => {
